requirements.txt
contourpy==1.0.6
customtkinter==4.6.3
cycler==0.11.0
darkdetect==0.7.1
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
seaborn==0.12.1
six==1.16.0
